+++ documentation/ChangeLog.md
@@ -2,6 +2,8 @@
 [Latest features on top]
 
 ### (NextVersion)
+ * Removed preference "Show error dialog if #DEAMON_TOOL failures occur" as it's no longer necessary.
+   * Content Assist will now distinguish between setup errors (errors that the user can fix, which are related to configuration and not the source) and other tool errors. For setup errors, a dialog will always be shown. For the other errors, if Content Assist was invoked manually, a dialog will also be shown, but if it was activated automatically the error will only be displayed in the editor status line. 
  * Fixed: Eclipse variables not being substituted in environment-variable values for launch configurations.
 
  * Fixed: Progress dialog never showing up when "Ok" pressed in LANG_NAME project properties page.